feat: move src/shared/resources to shared/constants/field
the old place was confusing. The exported elements are basically constants
