+++ shared/constants/field/myinfo/myinfo-occupations.ts
